/examples/lib moved to /lib
