+++ bindings/wasm/examples/package.json
@@ -5,9 +5,7 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "npm run build:vite && npm run build:transpile",
-    "build:vite": "tsc --noEmit && vite build",
-    "build:transpile": "npx tsc -p lib --outDir built/lib",
+    "build": "tsc --noEmit && vite build",
     "build:watch": "vite build --watch",
     "preview": "vite preview",
     "test": "vitest run",
@@ -15,22 +13,20 @@
     "test:dev": "vitest"
   },
   "dependencies": {
-    "@gltf-transform/core": "^3.8.0",
-    "@gltf-transform/extensions": "^3.8.0",
-    "@gltf-transform/functions": "^3.8.0",
+    "@gltf-transform/core": "^4.2.0",
+    "@gltf-transform/extensions": "^4.2.0",
+    "@gltf-transform/functions": "^4.2.0",
     "@jscadui/3mf-export": "^0.5.0",
-    "@types/three": "^0.164.0",
-    "commander": "^13.1.0",
-    "fflate": "^0.8.0",
-    "gl-matrix": "^3.4.3",
+    "@types/three": "^0.179.0",
+    "gl-matrix": "^3.4.4",
     "simple-dropzone": "0.8.3",
-    "three": "0.164.1"
+    "three": "0.179.1"
   },
   "devDependencies": {
-    "@types/node": "^22.10.2",
-    "@vitest/ui": "^1.6.0",
-    "typescript": "^5.7.0",
-    "vite": "^4.5.0",
-    "vitest": "^1.6.0"
+    "@types/node": "^24.2.1",
+    "@vitest/ui": "^3.2.4",
+    "typescript": "^5.9.2",
+    "vite": "^6.3.5",
+    "vitest": "^3.2.4"
   }
 }
