+++ bindings/wasm/lib/.gitignore
@@ -1,2 +1,3 @@
 *.js
 *.js.map
+*.d.ts
