+++ bindings/wasm/lib/manifold-gltf.ts
